Draft board election

+++ voting-microservice/src/main/java/voting/db/repos/ElectionRepository.java
@@ -12,7 +12,7 @@ public interface ElectionRepository extends JpaRepository<Election, Integer> {
 
     Optional<Election> findByElectionId(Integer electionId);
 
-    boolean existsByHoaIdAndName(int hoaId, String name);
+    boolean existsByHoaIdAndName(long hoaId, String name);
 
     @Query("SELECT u FROM BoardElection u WHERE u.hoaId = ?1 AND u.status LIKE 'ongoing'")
     Optional<Election> getBoardElectionByHoaId(int hoaId);
